Extract analysis metadata mutation executor
